Reuse loaded PIL Image for JPEG pixel analysis + DB pool fixes (v2.6.14)
Eliminate redundant Image.open() in JPEG pixel analysis by passing the
already-loaded PIL Image from the caller. Previously every JPEG was
opened 3 times (verify, load, pixel analysis), causing cumulative memory
growth that killed the worker after ~700 files.

Also adds pool_reset_on_return='rollback' and db.engine.dispose() on
DatabaseError to recover from corrupted connections after worker crashes.
