+++ CHANGELOG.MD
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
-## [2.6.13] - 2026-04-04
+## [2.6.14] - 2026-04-04
 
 ### Added
 
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **Fix JPEG pixel analysis crashing celery worker on large files**: Added file size guard (skip files > 10MB), image dimensions guard (skip images > 30MP), and timeout (30s) to prevent OOM kills and hangs when scanning large DSLR photos.
 - **Fix corrupted PostgreSQL connections after worker OOM kill**: Added `pool_reset_on_return='rollback'` to SQLAlchemy engine options and `db.engine.dispose()` on DatabaseError/OperationalError to force fresh connection pool after crashes. Prevents `psycopg2.DatabaseError: error with status PGRES_TUPLES_OK` from blocking all subsequent scans.
+- **Fix JPEG pixel analysis memory accumulation killing celery worker**: Refactored `_check_jpeg_pixel_corruption()` to accept the already-loaded PIL Image object instead of reopening the file. Previously every JPEG was opened three times (verify, load, pixel analysis), causing cumulative memory growth that killed the worker after ~700 files. Now reuses the existing loaded image with zero additional file I/O.
 - **Fix "0 of 0 files" progress display during scanning phase**: The scan status API returned `estimated_total=0` during the gap between phase transition and chunk counting. Added explicit Redis update after DB commit and a fallback to `phase_total` in the API when `estimated_total` is 0.
 - **Filter black frame false positives from freeze detection**: Freeze events that overlap with black sections (scene transitions, studio logos, end credits) are now filtered out. Uses FFmpeg `blackdetect` filter chained with `freezedetect` in a single decode pass.
 - **Fix scan stuck as active after completion**: Two issues -- (1) `_mark_scan_completed()` used raw SQL UPDATE but the ORM identity map still held stale `is_active=True`; when `_create_scan_report()` committed, it overwrote the completed state. Added `db.session.expire_all()` after the raw SQL commit. (2) Recovery endpoint and new scans didn't clear stale Redis progress keys, causing the API to return contradictory state (running + completed). Added `clear_scan_progress_redis()` to recovery and scan start.
