+++ tests/test_media_checker.py
@@ -5,11 +5,8 @@ Tests for PixelProbe media checker core functionality
 import pytest
 import os
 import subprocess
-import tempfile
-import time
 import threading
 from unittest.mock import Mock, patch, MagicMock
-from PIL import Image
 
 from pixelprobe.media_checker import PixelProbe
 
@@ -672,160 +669,3 @@ class TestVideoFreezeDetection:
         assert any('Filtered 1 of 2' in line for line in scan_output)
 
 
-class TestJpegPixelCorruption:
-    """Tests for JPEG pixel corruption detection."""
-
-    def test_jpeg_pixel_corruption_sustained_chaos(self):
-        """Detect corruption via sustained chaotic rows (rainbow garbage)."""
-        checker = PixelProbe()
-
-        # Create a 200x200 image with alternating color bands at the bottom
-        # simulating rainbow garbage -- each row drastically different from previous
-        # Using PNG to avoid JPEG compression smoothing the test signal
-        img = Image.new('RGB', (200, 200), (128, 128, 128))
-        pixels = img.load()
-        colors = [(255, 0, 0), (0, 0, 255), (0, 255, 0), (255, 255, 0)]
-        for y in range(160, 200):
-            c = colors[y % len(colors)]
-            for x in range(200):
-                pixels[x, y] = c
-
-        with tempfile.NamedTemporaryFile(suffix='.png', delete=False) as f:
-            img.save(f, 'PNG')
-            tmp_path = f.name
-
-        try:
-            is_corrupted, details, output = checker._check_jpeg_pixel_corruption(tmp_path)
-            assert is_corrupted is True
-            assert any('sustained chaos' in d for d in details)
-        finally:
-            os.unlink(tmp_path)
-
-    def test_jpeg_pixel_corruption_solid_fill(self):
-        """Detect corruption via solid fill streak preceded by chaos."""
-        checker = PixelProbe()
-
-        # Simulate real corruption: normal image -> chaotic garbage -> solid fill to EOF
-        img = Image.new('RGB', (200, 200), (100, 150, 200))
-        pixels = img.load()
-        # Rows 110-119: alternating solid colors (high inter-row jump averages)
-        chaos_colors = [(255, 0, 0), (0, 0, 255), (0, 255, 0), (255, 255, 0)]
-        for y in range(110, 120):
-            c = chaos_colors[y % len(chaos_colors)]
-            for x in range(200):
-                pixels[x, y] = c
-        # Rows 120-199: solid decoder fill to EOF
-        for y in range(120, 200):
-            for x in range(200):
-                pixels[x, y] = (128, 128, 128)
-
-        with tempfile.NamedTemporaryFile(suffix='.png', delete=False) as f:
-            img.save(f, 'PNG')
-            tmp_path = f.name
-
-        try:
-            is_corrupted, details, output = checker._check_jpeg_pixel_corruption(tmp_path)
-            assert is_corrupted is True
-            assert any('solid fill' in d for d in details)
-        finally:
-            os.unlink(tmp_path)
-
-    def test_jpeg_high_contrast_no_false_positive(self):
-        """YouTube-style thumbnail with sharp text boundary should not trigger."""
-        checker = PixelProbe()
-
-        # Simulate a thumbnail: dark top bar, then varied photo content below
-        # Single sharp boundary but no sustained chaos, no solid fill to bottom
-        img = Image.new('RGB', (200, 200))
-        pixels = img.load()
-        for y in range(0, 50):
-            for x in range(200):
-                pixels[x, y] = (10, 10, 10)  # dark bar
-        for y in range(50, 200):
-            for x in range(200):
-                pixels[x, y] = ((x * 3 + y) % 256, (y * 2) % 256, (x + y * 3) % 256)
-
-        with tempfile.NamedTemporaryFile(suffix='.png', delete=False) as f:
-            img.save(f, 'PNG')
-            tmp_path = f.name
-
-        try:
-            is_corrupted, details, output = checker._check_jpeg_pixel_corruption(tmp_path)
-            assert is_corrupted is False
-            assert any('PASSED' in line for line in output)
-        finally:
-            os.unlink(tmp_path)
-
-    def test_jpeg_clean_no_false_positive(self):
-        """Clean JPEG should not trigger corruption."""
-        checker = PixelProbe()
-
-        # Create a gradient image - smooth transitions, no sudden jumps
-        img = Image.new('RGB', (200, 200))
-        pixels = img.load()
-        for y in range(200):
-            for x in range(200):
-                pixels[x, y] = (x % 256, y % 256, (x + y) % 256)
-
-        with tempfile.NamedTemporaryFile(suffix='.jpg', delete=False) as f:
-            img.save(f, 'JPEG', quality=95)
-            tmp_path = f.name
-
-        try:
-            is_corrupted, details, output = checker._check_jpeg_pixel_corruption(tmp_path)
-            assert is_corrupted is False
-            assert len(details) == 0
-            assert any('PASSED' in line for line in output)
-        finally:
-            os.unlink(tmp_path)
-
-    def test_jpeg_pixel_corruption_skipped_for_png(self):
-        """Pixel corruption check should only run for JPEG files."""
-        checker = PixelProbe()
-
-        # The _check_image_corruption method gates on is_jpeg, so we verify
-        # the method itself works but the gate is what matters.
-        # Create a gradient PNG and verify the check still returns clean when called directly
-        img = Image.new('RGB', (100, 100))
-        pixels = img.load()
-        for y in range(100):
-            for x in range(100):
-                pixels[x, y] = (x * 2 % 256, y * 2 % 256, (x + y) % 256)
-        with tempfile.NamedTemporaryFile(suffix='.png', delete=False) as f:
-            img.save(f, 'PNG')
-            tmp_path = f.name
-
-        try:
-            # Even if called directly on a PNG, it should not false-positive
-            is_corrupted, details, output = checker._check_jpeg_pixel_corruption(tmp_path)
-            assert is_corrupted is False
-        finally:
-            os.unlink(tmp_path)
-
-    def test_jpeg_pixel_analysis_skipped_large_dimensions(self):
-        """High-resolution images should be skipped to avoid OOM."""
-        checker = PixelProbe()
-
-        # Create a small file but mock Image.open to report large dimensions
-        img = Image.new('RGB', (100, 100), (128, 128, 128))
-        with tempfile.NamedTemporaryFile(suffix='.png', delete=False) as f:
-            img.save(f, 'PNG')
-            tmp_path = f.name
-
-        try:
-            class FakeImg:
-                mode = 'RGB'
-                size = (8000, 6000)  # 48MP -- exceeds 30MP guard
-                def __enter__(self):
-                    return self
-                def __exit__(self, *args):
-                    pass
-                def convert(self, mode):
-                    return self
-
-            with patch('pixelprobe.media_checker.Image.open', return_value=FakeImg()):
-                is_corrupted, details, output = checker._check_jpeg_pixel_corruption(tmp_path)
-                assert is_corrupted is False
-                assert any('too large' in line for line in output)
-        finally:
-            os.unlink(tmp_path)
